kinship-ai-rewoo/unstructureddb_mcpserver/requirements.txt
# Core dependencies for MCP implementation (using latest available versions)
mcp
langchain-mcp-adapters
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
jsonschema>=4.18.0

# For UnstructuredDB tool
pinecone>=5.0.0
google-generativeai>=0.7.0

fastmcp

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.31.0 
uvicorn